Versao Final Teste

+++ app/src/main/java/appocorrencias/com/appocorrencias/ListView/ItemFeedOcorrencias.java
@@ -62,7 +62,8 @@ public class ItemFeedOcorrencias extends AppCompatActivity {
     ViewPager viewPager;
     AdapterCustomSwiper adapterCustomSwiper;
     private static ProcessaSocket processa = new ProcessaSocket();
-    String idOcorrencia, descricao, rua, bairro, uf, cidade, data, tipo, CPF, Nome, BairroCli, convComentario, CPFOcorrencia, tela;
+    String idOcorrencia, descricao, rua, bairro, uf, cidade, data, tipo, CPF, Nome, BairroCli, convComentario, CPFOcorrencia, tela, Ip, telaBusca;
+    int Porta;
     ListView listaComentarios;
 
     @Override
@@ -79,7 +80,6 @@ public class ItemFeedOcorrencias extends AppCompatActivity {
         TextView Tv_Cidade_UF = (TextView) findViewById(R.id.txtCidadeUFUsuario);
         TextView Tv_Rua_Bairro = (TextView) findViewById(R.id.txtRuaBairroUsuario);
         Button btnExcluir = (Button) findViewById(R.id.btnExcluir);
-        ImageButton btnDelComent = (ImageButton) findViewById(R.id.btnDelComent);
 
         Intent intent = getIntent();
 
@@ -91,6 +91,9 @@ public class ItemFeedOcorrencias extends AppCompatActivity {
         Nome = dados.getString("nome").toString();
         BairroCli = dados.getString("bairro").toString();
         tela = dados.getString("tela").toString();
+        telaBusca = dados.getString("telaBusca").toString();
+        Ip = dados.getString("ip").toString();
+        Porta = dados.getInt("porta");
 
 
         descricao = getDescricaoNr(idOcorrencia);
@@ -102,7 +105,7 @@ public class ItemFeedOcorrencias extends AppCompatActivity {
         tipo = getTipoNr(idOcorrencia);
         CPFOcorrencia = ArrayOcorrenciasRegistradas.getCPFNr(idOcorrencia);
 
-        if (CPFOcorrencia.equals(CPF)) {
+        if (CPFOcorrencia.equals(CPF) || CPF.equals("33333333333")) {
             btnExcluir.setVisibility(View.VISIBLE);
         }
 
@@ -157,7 +160,7 @@ public class ItemFeedOcorrencias extends AppCompatActivity {
         viewPager.setAdapter(adapterCustomSwiper);
     }
 
-    public static String evBuscarImagens(String IDOcorrencia, String tipo) throws IOException {
+    public static String evBuscarImagens(String IDOcorrencia, String tipo, String IpServer, int PortaServer) throws IOException {
 
         String BuscarImagensOcorrencia;
 
@@ -166,10 +169,6 @@ public class ItemFeedOcorrencias extends AppCompatActivity {
         } else {
             BuscarImagensOcorrencia = "BuscarImagensOcorrencia " + IDOcorrencia;
         }
-
-        String ip_conexao = "192.168.1.17"; /*// "52.34.140.131"; "172.20.10.3";*/
-        int porta_conexao = 2222; //63200;
-
 
         String str = null;
         OutputStream canalSaida = null;
@@ -186,7 +185,7 @@ public class ItemFeedOcorrencias extends AppCompatActivity {
         byte[] byteFinal = concat(byteTamanhoPct, TamanhoEDados);
 
         int millisecondsTimeOut = 5000;
-        InetSocketAddress adress = new InetSocketAddress(ip_conexao, porta_conexao);
+        InetSocketAddress adress = new InetSocketAddress(IpServer, PortaServer);
 
         try {
             cliente2.connect(adress, millisecondsTimeOut);
@@ -225,22 +224,21 @@ public class ItemFeedOcorrencias extends AppCompatActivity {
     public void evExcluirOcorrencia(View view) throws IOException {
 
         String ExcluirOcorrencia = "ExcluirOcorrencia " + idOcorrencia;
-        String retornoExclusao = processa.cadastrar1_no_server(ExcluirOcorrencia);
+        String retornoExclusao = processa.primeiroCadastroNoServidor(ExcluirOcorrencia, Ip, Porta);
 
         if (retornoExclusao.equals("true")) {
 
-            deleteAllArray();
+            ArrayOcorrenciasRegistradas.deleteAllArray();
 
             String BuscarOcorrenciasRegistradas = "BuscarOcorrenciasRegistradas" + " " + CPF;
 
             Toast.makeText(this, "Minhas Ocorrencias Registradas ", Toast.LENGTH_SHORT).show();
 
             ArrayImagensPerfilComentarios.deleteBitmap();
-            String retorno = ProcessaSocket.buscar_dados_imagens_server(BuscarOcorrenciasRegistradas);
+            String retorno = ProcessaSocket.buscarDadosImagensServer(BuscarOcorrenciasRegistradas, Ip, Porta);
 
             if (retorno.equals("false")) {
                 Toast.makeText(this, "Não há ocorrencias cadastradas", Toast.LENGTH_SHORT).show();
-                setContentView(R.layout.activity_listar_ocorrencias);
                 Intent cliente = new Intent(this, ListarOcorrencias.class);
                 Bundle bundle = new Bundle();
                 bundle.putString("nome", Nome);
@@ -275,7 +273,6 @@ public class ItemFeedOcorrencias extends AppCompatActivity {
                 }
                 Toast.makeText(this, "Mostrando suas Ocorrencias ", Toast.LENGTH_SHORT).show();
 
-                setContentView(R.layout.activity_listar_ocorrencias);
                 Intent cliente = new Intent(this, ListarOcorrencias.class);
 
                 Bundle bundle = new Bundle();
@@ -302,11 +299,8 @@ public class ItemFeedOcorrencias extends AppCompatActivity {
 
             Random random = new Random();
             int x = random.nextInt(101);
-            String NrAleatorio = Integer.toString(x);
             String BuscaId = "IDcomentario teste";
-            String IDserver = processa.cadastrar1_no_server(BuscaId);
-            //String IDComentario = IDserver + NrAleatorio;
-
+            String IDserver = processa.primeiroCadastroNoServidor(BuscaId, Ip, Porta);
 
             String ArrayNome[] = Nome.split(" ");
             String Apelido = ArrayNome[0];
@@ -324,9 +318,7 @@ public class ItemFeedOcorrencias extends AppCompatActivity {
             SimpleDateFormat hora2 = new SimpleDateFormat("HH:mm:h");
             String Hora = hora2.format(dtI.getTime());
 
-
-            String retorno = processa.cadastrar_Comentario(IDserver, idOcorrencia, CPF, Data, Hora, Apelido, convComentario);
-
+            String retorno = processa.cadastrar_Comentario(IDserver, idOcorrencia, CPF, Data, Hora, Apelido, convComentario, Ip, Porta);
 
             txtComentario.setText(null);
 
@@ -342,7 +334,7 @@ public class ItemFeedOcorrencias extends AppCompatActivity {
                     Toast.makeText(this, "Comantario Salvo com sucesso", Toast.LENGTH_SHORT).show();
 
                     deleteAllArrayComentarios();
-                    evBuscarComentario(idOcorrencia);
+                    evBuscarComentario(idOcorrencia, Ip, Porta);
 
                     listaComentarios = (ListView) findViewById(R.id.list_comentarios);
                     ArrayList<DadosComentarios> listadecomentarios = getListaComentarios();
@@ -354,16 +346,15 @@ public class ItemFeedOcorrencias extends AppCompatActivity {
                 }
             }
         }
-
     }
 
 
-    public static String evBuscarComentario(String IDOcorrencia) throws IOException {
+    public static String evBuscarComentario(String IDOcorrencia, String IpServer, int PortaServer) throws IOException {
 
         String BuscarComentariosRegistrados = "BuscarComentariosRegistrados " + IDOcorrencia;
 
         ArrayImagensPerfilComentarios.deleteBitmap();
-        String retorno = ProcessaSocket.buscar_dados_imagens_server(BuscarComentariosRegistrados);
+        String retorno = ProcessaSocket.buscarDadosImagensServer(BuscarComentariosRegistrados, IpServer, PortaServer);
 
         if (retorno.equals("erro")) {
             return "erro";
@@ -394,7 +385,6 @@ public class ItemFeedOcorrencias extends AppCompatActivity {
 
                     ArrayComentariosRegistrados.adicionar(dado);
 
-
                 }
             }
         }
@@ -402,14 +392,14 @@ public class ItemFeedOcorrencias extends AppCompatActivity {
     }
 
 
-    public void evOcorrenciasInformadasItem(String CPF) throws IOException {
+    public void evOcorrenciasInformadasItem(String CPF, String Ip, int Porta) throws IOException {
 
         deleteAllArray();
 
         String BuscarOcorrenciasRegistradas = "BuscarOcorrenciasRegistradas" + " " + CPF;
 
         ArrayImagensPerfilComentarios.deleteBitmap();
-        String retorno = ProcessaSocket.buscar_dados_imagens_server(BuscarOcorrenciasRegistradas);
+        String retorno = ProcessaSocket.buscarDadosImagensServer(BuscarOcorrenciasRegistradas, Ip, Porta);
 
         if (retorno.equals("false")) {
 
@@ -445,10 +435,8 @@ public class ItemFeedOcorrencias extends AppCompatActivity {
             }
 
             Toast.makeText(this, "Mostrando suas Ocorrencias ", Toast.LENGTH_SHORT).show();
-
         }
     }
-
 
     public void onBackPressed() {
         super.onBackPressed();
@@ -460,7 +448,7 @@ public class ItemFeedOcorrencias extends AppCompatActivity {
 
             deleteAllArray();
             try {
-                evOcorrenciasInformadasItem(CPF);
+                evOcorrenciasInformadasItem(CPF, Ip, Porta);
             } catch (IOException e) {
                 e.printStackTrace();
             }
@@ -469,18 +457,20 @@ public class ItemFeedOcorrencias extends AppCompatActivity {
             bundle.putString("nome", Nome);
             bundle.putString("cpf", CPF);
             bundle.putString("bairro", BairroCli);
+            bundle.putString("ip", Ip);
+            bundle.putInt("porta", Porta);
 
             cliente.putExtras(bundle);
             this.startActivity(cliente);
             this.finish();
 
         } else {
-            if (tela.equals("Cliente")) {
+            if (tela.equals("Cliente") && (telaBusca.equals("Busca") == false)) {
 
                 deleteAllArray();
 
                 try {
-                    evBuscarOcorrenciasBairro(BairroCli);
+                    evBuscarOcorrenciasBairro(BairroCli, Ip, Porta);
                 } catch (IOException e) {
                     e.printStackTrace();
                 }
@@ -491,27 +481,48 @@ public class ItemFeedOcorrencias extends AppCompatActivity {
                 bundle.putString("nome", Nome);
                 bundle.putString("cpf", CPF);
                 bundle.putString("bairro", BairroCli);
+                bundle.putString("ip", Ip);
+                bundle.putInt("porta", Porta);
 
                 cliente.putExtras(bundle);
                 this.startActivity(cliente);
                 this.finish();
 
             } else {
-                Intent cliente = new Intent(this, BuscarOcorrencias.class);
+                if (tela.equals("Cliente") && telaBusca.equals("Busca")) {
+                    Intent cliente = new Intent(this, BuscarOcorrencias.class);
 
-                Bundle bundle = new Bundle();
-                bundle.putString("nome", Nome);
-                bundle.putString("cpf", CPF);
-                bundle.putString("bairro", BairroCli);
-                bundle.putString("tela", tela);
+                    Bundle bundle = new Bundle();
+                    bundle.putString("nome", Nome);
+                    bundle.putString("cpf", CPF);
+                    bundle.putString("bairro", BairroCli);
+                    bundle.putString("tela", tela);
+                    bundle.putString("telaBusca", telaBusca);
+                    bundle.putString("ip", Ip);
+                    bundle.putInt("porta", Porta);
 
-                cliente.putExtras(bundle);
-                this.startActivity(cliente);
-                this.finish();
+                    cliente.putExtras(bundle);
+                    this.startActivity(cliente);
+                    this.finish();
+                } else {
+                    if (tela.equals("Adm") && telaBusca.equals("Busca")) {
+                        Intent cliente = new Intent(this, BuscarOcorrencias.class);
 
+                        Bundle bundle = new Bundle();
+                        bundle.putString("nome", Nome);
+                        bundle.putString("cpf", CPF);
+                        bundle.putString("bairro", BairroCli);
+                        bundle.putString("tela", tela);
+                        bundle.putString("telaBusca", telaBusca);
+                        bundle.putString("ip", Ip);
+                        bundle.putInt("porta", Porta);
 
+                        cliente.putExtras(bundle);
+                        this.startActivity(cliente);
+                        this.finish();
+                    }
+                }
             }
         }
     }
-
 }
